pyproject.toml
[project]
name = "local-ai"
version = "0.1.0"
description = "Cleaned up AI assistant project"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "duckduckgo-search>=8.1.1",
    "google-genai>=1.61.0",
    "mcp>=1.26.0",
    "numpy>=2.3.5",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sounddevice>=0.5.5",
    "websockets>=14.1",
    "ytmusicapi>=1.11.5",
]
[tool.setuptools]
packages = ["chip"]
